Pass /no-preprocess flag to llvm-rc 16+
Since LLVM 16, llvm-rc supports doing the preprocessing itself. However,
doing so requires having `clang` in the PATH, which some more complex
toolchain may not necessarily have (for instance, for windows, you may
have `clang-cl` but no `clang` in PATH, or you may have a
target-prefixed clang like `x86_64-pc-windows-msvc-clang`).

If llvm-rc fails to find clang, it will error out with a message saying
it failed to find the program in PATH.

To avoid these issues, we continue doing the preprocessing manually, and
pass the /no-preprocess flag to llvm-rc.